Add function to subscribe to notifications given a tracking number
- Refactor models to different files
- Move old fedex stuff to fedex_old.go
- Refactor tests to be neater, read from creds file
