+++ fedex.go
@@ -1,6 +1,6 @@
 // History: Nov 20 13 tcolar Creation
 
-// fedex provides access to (some) FedEx Soap API's and unmarshall answers into Go structures
+// Package fedex provides access to (some) FedEx Soap API's and unmarshal answers into Go structures
 package fedex
 
 import (
@@ -13,8 +13,8 @@ import (
 	"github.com/happyreturns/fedex/models"
 )
 
+// Convenience constants for standard Fedex API URLs
 const (
-	// Convenience constants for standard Fedex API url's
 	FedexAPIURL               = "https://ws.fedex.com:443/web-services"
 	FedexAPITestURL           = "https://wsbeta.fedex.com:443/web-services"
 	CarrierCodeExpress        = "FDXE"
@@ -37,64 +37,96 @@ type Fedex struct {
 	FedexURL string
 }
 
-func (f Fedex) wrapSoapRequest(body string, namespace string) string {
-	return fmt.Sprintf(`
-	<soapenv:Envelope xmlns:soapenv="http://schemas.xmlsoap.org/soap/envelope/" xmlns:q0="%s">
-		<soapenv:Body>
-			%s
-		</soapenv:Body>
-	</soapenv:Envelope>
-	`, namespace, body)
-}
+// TrackByNumber returns tracking info for a specific Fedex tracking number
+func (f Fedex) TrackByNumber(carrierCode string, trackingNo string) (*models.TrackReply, error) {
 
-func (f Fedex) soapCreds(serviceID, majorVersion string) string {
-	return fmt.Sprintf(`
-		<q0:WebAuthenticationDetail>
-			<q0:UserCredential>
-				<q0:Key>%s</q0:Key>
-				<q0:Password>%s</q0:Password>
-			</q0:UserCredential>
-		</q0:WebAuthenticationDetail>
-		<q0:ClientDetail>
-			<q0:AccountNumber>%s</q0:AccountNumber>
-			<q0:MeterNumber>%s</q0:MeterNumber>
-		</q0:ClientDetail>
-		<q0:Version>
-			<q0:ServiceId>%s</q0:ServiceId>
-			<q0:Major>%s</q0:Major>
-			<q0:Intermediate>0</q0:Intermediate>
-			<q0:Minor>0</q0:Minor>
-		</q0:Version>
-	`, f.Key, f.Password, f.Account, f.Meter, serviceID, majorVersion)
-}
+	request := f.trackByNumberRequest(carrierCode, trackingNo)
+	response := &models.TrackResponseEnvelope{}
 
-// TrackByShipperRef : Return tracking info for a specific shipper reference
-// ShipperRef is usually an order ID or other unique identifier
-// ShipperAccountNumber is the Fedex account number of the shipper
-func (f Fedex) TrackByShipperRef(carrierCode string, shipperRef string,
-	shipperAccountNumber string) (reply models.TrackReply, err error) {
-	reqXML := soapRefTracking(f, carrierCode, shipperRef, shipperAccountNumber)
-	content, err := f.PostXML(f.FedexURL+"/trck", reqXML)
+	err := f.makeRequestAndUnmarshalResponse("/trck", request, response)
 	if err != nil {
-		return reply, err
+		return nil, fmt.Errorf("make track request and unmarshal: %s", err)
 	}
-	return f.parseTrackReply(content)
+	return &response.Reply, nil
 }
 
-// TrackByPo : Returns tracking info for a specific Purchase Order (often the OrderId)
-// Note that Fedex requires the Destination Postal Code & country
-//   to match when making PO queries
-func (f Fedex) TrackByPo(carrierCode string, po string, postalCode string,
-	countryCode string) (reply models.TrackReply, err error) {
-	reqXML := soapPoTracking(f, carrierCode, po, postalCode, countryCode)
-	content, err := f.PostXML(f.FedexURL+"/trck", reqXML)
+// ShipGround creates a ground shipment
+func (f Fedex) ShipGround(shipment *models.Shipment) (*models.ProcessShipmentReply, error) {
+
+	request, err := f.shipmentRequest("FEDEX_GROUND", shipment)
 	if err != nil {
-		return reply, err
+		return nil, fmt.Errorf("create shipment request: %s", err)
 	}
-	return f.parseTrackReply(content)
+
+	response := &models.ShipResponseEnvelope{}
+	err = f.makeRequestAndUnmarshalResponse("/ship/v23", request, response)
+	if err != nil {
+		return nil, fmt.Errorf("make ship ground request and unmarshal: %s", err)
+	}
+	return &response.Reply, nil
 }
 
-func (f Fedex) makeRequestAndUnmarshal(url string, request models.Envelope,
+// ShipSmartPost creates a Smart Post return shipment
+func (f Fedex) ShipSmartPost(shipment *models.Shipment) (*models.ProcessShipmentReply, error) {
+
+	request, err := f.shipmentRequest("SMART_POST", shipment)
+	if err != nil {
+		return nil, fmt.Errorf("create shipment request: %s", err)
+	}
+
+	response := &models.ShipResponseEnvelope{}
+	err = f.makeRequestAndUnmarshalResponse("/ship/v23", request, response)
+	if err != nil {
+		return nil, fmt.Errorf("make ship smart post request and unmarshal: %s", err)
+	}
+
+	return &response.Reply, nil
+}
+
+// Rate : Gets the estimated rates for a shipment
+func (f Fedex) Rate(fromAddress models.Address, toAddress models.Address,
+	fromContact models.Contact, toContact models.Contact) (*models.RateReply, error) {
+
+	request := f.rateRequest(fromAddress, toAddress, fromContact, toContact)
+	response := &models.RateResponseEnvelope{}
+
+	err := f.makeRequestAndUnmarshalResponse("/rate/v24", request, response)
+	if err != nil {
+		return nil, fmt.Errorf("make rate request and unmarshal: %s", err)
+	}
+
+	return &response.Reply, nil
+}
+
+// CreatePickup creates a pickup
+func (f Fedex) CreatePickup(pickupLocation models.PickupLocation, toAddress models.Address) (*models.CreatePickupReply, error) {
+
+	request := f.createPickupRequest(pickupLocation, toAddress)
+	response := &models.CreatePickupResponseEnvelope{}
+
+	err := f.makeRequestAndUnmarshalResponse("/pickup/v17", request, response)
+	if err != nil {
+		return nil, fmt.Errorf("make create pickup request and unmarshal: %s", err)
+	}
+
+	return &response.Reply, nil
+}
+
+// SendNotifications gets notifications sent to an email
+func (f Fedex) SendNotifications(trackingNo, email string) (*models.SendNotificationsReply, error) {
+
+	request := f.notificationsRequest(trackingNo, email)
+	response := &models.SendNotificationsResponseEnvelope{}
+
+	err := f.makeRequestAndUnmarshalResponse("/track/v16", request, response)
+	if err != nil {
+		return nil, fmt.Errorf("make send notifications request: %s", err)
+	}
+
+	return &response.Reply, nil
+}
+
+func (f Fedex) makeRequestAndUnmarshalResponse(url string, request models.Envelope,
 	response models.Response) error {
 	// Create request body
 	reqXML, err := xml.Marshal(request)
@@ -103,7 +135,7 @@ func (f Fedex) makeRequestAndUnmarshal(url string, request models.Envelope,
 	}
 
 	// Post XML
-	content, err := f.PostXML(f.FedexURL+url, string(reqXML))
+	content, err := f.postXML(f.FedexURL+url, string(reqXML))
 	if err != nil {
 		return fmt.Errorf("post xml: %s", err)
 	}
@@ -123,89 +155,8 @@ func (f Fedex) makeRequestAndUnmarshal(url string, request models.Envelope,
 	return nil
 }
 
-// TrackByNumber : Returns tracking info for a specific Fedex tracking number
-func (f Fedex) TrackByNumber(carrierCode string, trackingNo string) (*models.TrackReply, error) {
-	request := f.trackByNumberSOAPRequest(carrierCode, trackingNo)
-	response := &models.TrackResponseEnvelope{}
-
-	err := f.makeRequestAndUnmarshal("/trck", request, response)
-	if err != nil {
-		return nil, fmt.Errorf("make track request and unmarshal: %s", err)
-	}
-	return &response.Reply, nil
-}
-
-// ShipGround : Creates a ground shipment
-func (f Fedex) ShipGround(shipment *models.Shipment) (*models.ProcessShipmentReply, error) {
-
-	request, err := f.shipmentEnvelope("FEDEX_GROUND", shipment)
-	if err != nil {
-		return nil, fmt.Errorf("create shipment request: %s", err)
-	}
-
-	response := &models.ShipResponseEnvelope{}
-	err = f.makeRequestAndUnmarshal("/ship/v23", request, response)
-	if err != nil {
-		return nil, fmt.Errorf("make ship ground request and unmarshal: %s", err)
-	}
-	return &response.Reply, nil
-}
-
-// ShipSmartPost : Creates a Smart Post return shipment
-func (f Fedex) ShipSmartPost(shipment *models.Shipment) (*models.ProcessShipmentReply, error) {
-
-	request, err := f.shipmentEnvelope("SMART_POST", shipment)
-	if err != nil {
-		return nil, fmt.Errorf("create shipment request: %s", err)
-	}
-
-	response := &models.ShipResponseEnvelope{}
-	err = f.makeRequestAndUnmarshal("/ship/v23", request, response)
-	if err != nil {
-		return nil, fmt.Errorf("make ship smart post request and unmarshal: %s", err)
-	}
-	return &response.Reply, nil
-
-}
-
-// Rate : Gets the estimated rates for a shipment
-func (f Fedex) Rate(fromAddress models.Address, toAddress models.Address,
-	fromContact models.Contact, toContact models.Contact) (*models.RateReply, error) {
-
-	request := f.rateSOAPRequest(fromAddress, toAddress, fromContact, toContact)
-	response := &models.RateResponseEnvelope{}
-
-	err := f.makeRequestAndUnmarshal("/rate/v24", request, response)
-	if err != nil {
-		return nil, fmt.Errorf("make rate request and unmarshal: %s", err)
-	}
-	return &response.Reply, nil
-}
-
-// TODO
-func (f Fedex) CreatePickup(pickupLocation models.PickupLocation, toAddress models.Address) (*models.CreatePickupReply, error) {
-
-	request := f.createPickupRequest(pickupLocation, toAddress)
-	response := &models.CreatePickupResponseEnvelope{}
-
-	err := f.makeRequestAndUnmarshal("/pickup/v17", request, response)
-	if err != nil {
-		return nil, fmt.Errorf("make create pickup request and unmarshal: %s", err)
-	}
-	return &response.Reply, nil
-}
-
-// Unmarshal XML SOAP response into a TrackReply
-func (f Fedex) parseTrackReply(xmlResp []byte) (reply models.TrackReply, err error) {
-	data := struct {
-		Reply models.TrackReply `xml:"Body>TrackReply"`
-	}{}
-	err = xml.Unmarshal(xmlResp, &data)
-	return data.Reply, err
-}
-
-// Post Xml to Fedex API and return response
-func (f Fedex) PostXML(url string, xml string) (content []byte, err error) {
+// postXML to Fedex API and return response
+func (f Fedex) postXML(url string, xml string) (content []byte, err error) {
 	resp, err := http.Post(url, "text/xml", strings.NewReader(xml))
 	if err != nil {
 		return content, err
