+++ models/models.go
@@ -1,13 +1,6 @@
 package models
 
-import (
-	"errors"
-	"fmt"
-	"time"
-)
-
-const notificationSeverityError = "ERROR"
-const notificationSeveritySuccess = "SUCCESS"
+import "time"
 
 // Shipment is convenience struct that has fields for creating a shipment (not part of FedEx API)
 type Shipment struct {
@@ -18,109 +11,143 @@ type Shipment struct {
 	NotificationEmail string
 }
 
-// Envelope is the soap wrapper for all requests
-type Envelope struct {
-	XMLName   string      `xml:"soapenv:Envelope"`
-	Body      interface{} `xml:"soapenv:Body"`
-	Soapenv   string      `xml:"xmlns:soapenv,attr"`
-	Namespace string      `xml:"xmlns:q0,attr"`
+type Address struct {
+	StreetLines         []string `xml:"q0:StreetLines"`
+	City                string   `xml:"q0:City"`
+	StateOrProvinceCode string   `xml:"q0:StateOrProvinceCode"`
+	PostalCode          string   `xml:"q0:PostalCode"`
+	CountryCode         string   `xml:"q0:CountryCode"`
+	// CountryName         string   `xml:"q0:CountryName"`
+	Residential Bool `xml:"q0:Residential"`
 }
 
-type Response interface {
-	Error() error
+type AdvanceNotificationDetail struct {
+	EstimatedTimeOfArrival Timestamp
+	Reason                 string
+	Status                 string
+	StatusDescription      string
+	StatusTime             Timestamp
 }
 
-type TrackResponseEnvelope struct {
-	Reply TrackReply `xml:"Body>TrackReply"`
+type AncillaryDetail struct {
+	Reason            string
+	ReasonDescription string
 }
 
-func (t *TrackResponseEnvelope) Error() error {
-	// TrackResponses are odd in that for invalid tracking numbers, the Reply
-	// doesn't say it errored, even though the Reply.CompletedTrackDetails does
-
-	// Error if Reply has error
-	err := t.Reply.Error()
-	if err != nil {
-		return fmt.Errorf("track reply error: %s", err)
-	}
-
-	// Error if CompletedTrackDetails has error
-	for _, completedTrackDetail := range t.Reply.CompletedTrackDetails {
-		for _, trackDetail := range completedTrackDetail.TrackDetails {
-			if trackDetail.Notification.Severity == notificationSeverityError {
-				return fmt.Errorf("track detail error: %s", trackDetail.Notification.Message)
-			}
-		}
-	}
-
-	return nil
+type Charge struct {
+	Currency string
+	Amount   string
 }
 
-type ShipResponseEnvelope struct {
-	Reply ProcessShipmentReply `xml:"Body>ProcessShipmentReply"`
+type CompletedPackageDetails struct {
+	SequenceNumber string
+	TrackingIds    []TrackingID
+	Label          Label
 }
 
-func (s *ShipResponseEnvelope) Error() error {
-	return s.Reply.Error()
+type CompletedShipmentDetail struct {
+	UsDomestic              string
+	CarrierCode             string
+	MasterTrackingId        TrackingID
+	ServiceTypeDescription  string
+	ServiceDescription      ServiceDescription
+	PackagingDescription    string
+	OperationalDetail       OperationalDetail
+	ShipmentRating          Rating
+	CompletedPackageDetails CompletedPackageDetails
 }
 
-type RateResponseEnvelope struct {
-	Reply RateReply `xml:"Body>RateReply"`
+type CompletedTrackDetail struct {
+	HighestSeverity  string
+	Notifications    []Notification
+	DuplicateWaybill bool
+	MoreData         bool
+	TrackDetails     []TrackDetail
 }
 
-func (r *RateResponseEnvelope) Error() error {
-	return r.Reply.Error()
+type Contact struct {
+	PersonName   string `xml:"q0:PersonName"`
+	CompanyName  string `xml:"q0:CompanyName"`
+	PhoneNumber  string `xml:"q0:PhoneNumber"`
+	EmailAddress string `xml:"q0:EMailAddress"`
 }
 
-type CreatePickupResponseEnvelope struct {
-	Reply CreatePickupReply `xml:"Body>CreatePickupReply"`
+type ContactAndAddress struct {
+	Contact Contact `xml:"q0:Contact"`
+	Address Address `xml:"q0:Address"`
 }
 
-func (c *CreatePickupResponseEnvelope) Error() error {
-	return c.Reply.Error()
+type ContentRecord struct {
+	PartNumber       string
+	ItemNumber       string
+	ReceivedQuantity int
+	Description      string
 }
 
-// Request has just the default auth fields on all requests
-type Request struct {
-	WebAuthenticationDetail WebAuthenticationDetail `xml:"q0:WebAuthenticationDetail"`
-	ClientDetail            ClientDetail            `xml:"q0:ClientDetail"`
-	// TransitionDetail        *TransactionDetail      `xml:"q0:TransitionDetail,omitempty"`
-	Version Version `xml:"q0:Version"`
+type CustomerReference struct {
+	CustomerReferenceType string `xml:"q0:CustomerReferenceType"`
+	Value                 string `xml:"q0:Value"`
 }
 
-type WebAuthenticationDetail struct {
-	UserCredential UserCredential `xml:"q0:UserCredential"`
+type DateOrTimestamp struct {
+	Type            string
+	DateOrTimestamp Timestamp
 }
 
-type ClientDetail struct {
-	AccountNumber string `xml:"q0:AccountNumber"`
-	MeterNumber   string `xml:"q0:MeterNumber"`
+type Destination struct {
+	StreetLines         []string
+	City                string
+	StateOrProvinceCode string
+	PostalCode          string
+	CountryCode         string
+	CountryName         string
+	Residential         bool
 }
 
-type UserCredential struct {
-	Key      string `xml:"q0:Key"`
-	Password string `xml:"q0:Password"`
+type Dimensions struct {
+	Length int    `xml:"q0:Length"`
+	Width  int    `xml:"q0:Width"`
+	Height int    `xml:"q0:Height"`
+	Units  string `xml:"q0:Units"`
 }
 
-type RateRequest struct {
-	Request
-	RequestedShipment RequestedShipment `xml:"q0:RequestedShipment"`
+type EmailDetail struct {
+	EmailAddress string `xml:"q0:EmailAddress"`
+	Name         string `xml:"q0:Name"`
 }
 
-type TrackRequest struct {
-	Request
-	SelectionDetails  SelectionDetails `xml:"q0:SelectionDetails"`
-	ProcessingOptions string           `xml:"q0:ProcessingOptions"`
+type Event struct {
+	Timestamp                  time.Time
+	EventType                  string
+	EventDescription           string
+	StatusExceptionCode        string
+	StatusExceptionDescription string
+	Address                    Address
+	ArrivalLocation            string
 }
 
-type CreatePickupRequest struct {
-	Request
-	OriginDetail         OriginDetail        `xml:"q0:OriginDetail"`
-	FreightPickupDetail  FreightPickupDetail `xml:"q0:FreightPickupDetail"`
-	PackageCount         int                 `xml:"q0:PackageCount"`
-	CarrierCode          string              `xml:"q0:CarrierCode"`
-	Remarks              string              `xml:"q0:Remarks"`
-	CommodityDescription string              `xml:"q0:CommodityDescription"`
+type EventNotification struct {
+	Role                string              `xml:"q0:Role"`
+	Events              []string            `xml:"q0:Events"`
+	NotificationDetail  NotificationDetail  `xml:"q0:NotificationDetail"`
+	FormatSpecification FormatSpecification `xml:"q0:FormatSpecification"`
+}
+
+type EventNotificationDetail struct {
+	AggregationType    string              `xml:"q0:AggregationType"`
+	PersonalMessage    string              `xml:"q0:PersonalMessage"`
+	EventNotifications []EventNotification `xml:"q0:EventNotifications"`
+}
+
+// type EventNotificationDetailTracking struct {
+// 	Role                string              `xml:"q0:Role"`
+// 	Events              []string            `xml:"q0:Events"`
+// 	NotificationDetail  NotificationDetail  `xml:"q0:NotificationDetail"`
+// 	FormatSpecification FormatSpecification `xml:"q0:FormatSpecification"`
+// }
+
+type FormatSpecification struct {
+	Type string `xml:"q0:Type"`
 }
 
 type FreightPickupDetail struct {
@@ -143,6 +170,56 @@ type FreightPickupLineItem struct {
 	Description        string  `xml:"q0:Description"`
 }
 
+type Identifier struct {
+	Type  string
+	Value string
+}
+
+type InformationNoteDetail struct {
+	Code        string
+	Description string
+}
+
+type Label struct {
+	Type                        string
+	ShippingDocumentDisposition string
+	ImageType                   string
+	Resolution                  string
+	CopiesToPrint               string
+	Parts                       []Part
+}
+
+type LabelSpecification struct {
+	LabelFormatType string `xml:"q0:LabelFormatType"`
+	ImageType       string `xml:"q0:ImageType"`
+}
+
+type Localization struct {
+	LanguageCode string `xml:"q0:LanguageCode"`
+}
+
+type Name struct {
+	Type     string
+	Encoding string
+	Value    string
+}
+
+type NotificationDetail struct {
+	NotificationType string       `xml:"q0:NotificationType"`
+	EmailDetail      EmailDetail  `xml:"q0:EmailDetail"`
+	Localization     Localization `xml:"q0:Localization"`
+}
+
+type OperationalDetail struct {
+	OriginLocationNumber            string
+	DestinationLocationNumber       string
+	TransitTime                     string
+	IneligibleForMoneyBackGuarantee string
+	DeliveryEligibilities           string
+	ServiceCode                     string
+	PackagingCode                   string
+}
+
 type OriginDetail struct {
 	UseAccountAddress       Bool           `xml:"q0:UseAccountAddress"`
 	PickupLocation          PickupLocation `xml:"q0:PickupLocation"`
@@ -153,33 +230,100 @@ type OriginDetail struct {
 	CompanyCloseTime        string         `xml:"q0:CompanyCloseTime"`
 }
 
-type PickupLocation struct {
-	Contact Contact `xml:"q0:Contact"`
-	Address Address `xml:"q0:Address"`
-}
-
-type SelectionDetails struct {
-	CarrierCode       string            `xml:"q0:CarrierCode"`
-	PackageIdentifier PackageIdentifier `xml:"q0:PackageIdentifier"`
-	// Destination           Destination
-	// ShipmentAccountNumber string
-}
-
 type PackageIdentifier struct {
 	Type  string `xml:"q0:Type"`
 	Value string `xml:"q0:Value"`
 }
 
-type Version struct {
-	ServiceID    string `xml:"q0:ServiceId"`
-	Major        int    `xml:"q0:Major"`
-	Intermediate int    `xml:"q0:Intermediate"`
-	Minor        int    `xml:"q0:Minor"`
+type Part struct {
+	DocumentPartSequenceNumber string
+	Image                      []byte
 }
 
-type ProcessShipmentRequest struct {
-	Request
-	RequestedShipment RequestedShipment `xml:"q0:RequestedShipment"`
+type Payment struct {
+	PaymentType string `xml:"q0:PaymentType"`
+	Payor       Payor  `xml:"q0:Payor"`
+}
+
+type Payor struct {
+	ResponsibleParty ResponsibleParty `xml:"q0:ResponsibleParty"`
+}
+
+type RateDetail struct {
+	RateType                         string
+	RateZone                         string
+	RatedWeightMethod                string
+	DimDivisor                       string
+	FuelSurchargePercent             string
+	TotalBillingWeight               Weight
+	TotalBaseCharge                  Charge
+	TotalFreightDiscounts            Charge
+	TotalNetFreight                  Charge
+	TotalSurcharges                  Charge
+	TotalNetFedExCharge              Charge
+	TotalTaxes                       Charge
+	TotalNetCharge                   Charge
+	NetCharge                        Charge
+	TotalRebates                     Charge
+	TotalDutiesAndTaxes              Charge
+	TotalAncillaryFeesAndTaxes       Charge
+	TotalDutiesTaxesAndFees          Charge
+	TotalNetChargeWithDutiesAndTaxes Charge
+	Surcharges                       []Surcharge
+}
+
+type RateReplyDetail struct {
+	ServiceType                     string
+	ServiceDescription              ServiceDescription
+	PackagingType                   string
+	DestinationAirportID            string `xml:"DestinationAirportId"`
+	IneligibleForMoneyBackGuarantee bool
+	SignatureOption                 string
+	ActualRateType                  string
+	RatedShipmentDetails            []Rating // TODO
+}
+
+type RatedPackage struct {
+	GroupNumber          string
+	EffectiveNetDiscount Charge
+	PackageRateDetail    RateDetail
+}
+
+type RatedShipmentDetail struct {
+	EffectiveNetDiscount Charge
+	ShipmentRateDetail   RateDetail
+	RatedPackages        []RatedPackage
+}
+
+type Rating struct {
+	ActualRateType       string
+	GroupNumber          string
+	EffectiveNetDiscount Charge
+
+	// For the shipping service, the rate details is an array, but for the rate service, it is not
+	ShipmentRateDetails []RateDetail
+	ShipmentRateDetail  RateDetail
+
+	RatedPackages []RatedPackage
+}
+
+type RecipientDetail struct {
+	NotificationEventsAvailable []string
+}
+
+type Reconciliation struct {
+	Status      string
+	Description string
+}
+
+type RequestedPackageLineItem struct {
+	SequenceNumber     int                 `xml:"q0:SequenceNumber"`
+	GroupPackageCount  int                 `xml:"q0:GroupPackageCount,omitempty"`
+	Weight             Weight              `xml:"q0:Weight"`
+	Dimensions         Dimensions          `xml:"q0:Dimensions"`
+	PhysicalPackaging  string              `xml:"q0:PhysicalPackaging"`
+	ItemDescription    string              `xml:"q0:ItemDescription"`
+	CustomerReferences []CustomerReference `xml:"q0:CustomerReferences"`
 }
 
 type RequestedShipment struct {
@@ -207,277 +351,39 @@ type RequestedShipment struct {
 	RequestedPackageLineItems []RequestedPackageLineItem `xml:"q0:RequestedPackageLineItems"`
 }
 
-type SpecialServicesRequested struct {
-	SpecialServiceTypes     []string                 `xml:"q0:SpecialServiceTypes,omitempty"`
-	EventNotificationDetail *EventNotificationDetail `xml:"q0:EventNotificationDetail,omitempty"`
-	ReturnShipmentDetail    *ReturnShipmentDetail    `xml:"q0:ReturnShipmentDetail,omitempty"`
-}
-
-type EventNotificationDetail struct {
-	AggregationType    string              `xml:"q0:AggregationType"`
-	PersonalMessage    string              `xml:"q0:PersonalMessage"`
-	EventNotifications []EventNotification `xml:"q0:EventNotifications"`
-}
-
-type EventNotification struct {
-	Role                string              `xml:"q0:Role"`
-	Events              []string            `xml:"q0:Events"`
-	NotificationDetail  NotificationDetail  `xml:"q0:NotificationDetail"`
-	FormatSpecification FormatSpecification `xml:"q0:FormatSpecification"`
-}
-
-type NotificationDetail struct {
-	NotificationType string       `xml:"q0:NotificationType"`
-	EmailDetail      EmailDetail  `xml:"q0:EmailDetail"`
-	Localization     Localization `xml:"q0:Localization"`
-}
-
-type Localization struct {
-	LanguageCode string `xml:"q0:LanguageCode"`
-}
-
-type EmailDetail struct {
-	EmailAddress string `xml:"q0:EmailAddress"`
-	Name         string `xml:"q0:Name"`
-}
-
-type FormatSpecification struct {
-	Type string `xml:"q0:Type"`
-}
-
-type NotificationFormatType struct {
+type ResponsibleParty struct {
+	AccountNumber string `xml:"q0:AccountNumber"`
 }
 
 type ReturnShipmentDetail struct {
 	ReturnType string `xml:"q0:ReturnType"`
 }
 
-type ShipmentManifestDetail struct {
-	ManifestReferenceType string `xml:"q0:ManifestReferenceType,omitempty"`
+type Package struct {
+	TrackingNumber                  string
+	TrackingNumberUniqueIdentifiers []string
+	CarrierCode                     string
+	ShipDate                        string
+	Destination                     Destination
+	RecipientDetails                []RecipientDetail
 }
 
-type SmartPostDetail struct {
-	Indicia              string `xml:"q0:Indicia"`
-	AncillaryEndorsement string `xml:"q0:AncillaryEndorsement"`
-	HubID                string `xml:"q0:HubId"`
-}
-type RequestedPackageLineItem struct {
-	SequenceNumber     int                 `xml:"q0:SequenceNumber"`
-	GroupPackageCount  int                 `xml:"q0:GroupPackageCount,omitempty"`
-	Weight             Weight              `xml:"q0:Weight"`
-	Dimensions         Dimensions          `xml:"q0:Dimensions"`
-	PhysicalPackaging  string              `xml:"q0:PhysicalPackaging"`
-	ItemDescription    string              `xml:"q0:ItemDescription"`
-	CustomerReferences []CustomerReference `xml:"q0:CustomerReferences"`
+type PickupLocation struct {
+	Contact Contact `xml:"q0:Contact"`
+	Address Address `xml:"q0:Address"`
 }
 
-type CustomerReference struct {
-	CustomerReferenceType string `xml:"q0:CustomerReferenceType"`
-	Value                 string `xml:"q0:Value"`
+type SelectionDetails struct {
+	CarrierCode       string            `xml:"q0:CarrierCode"`
+	PackageIdentifier PackageIdentifier `xml:"q0:PackageIdentifier"`
+	// Destination           Destination
+	// ShipmentAccountNumber string
 }
 
-type Weight struct {
-	Units string  `xml:"q0:Units"`
-	Value float64 `xml:"q0:Value"`
-}
-
-type Contact struct {
-	PersonName   string `xml:"q0:PersonName"`
-	CompanyName  string `xml:"q0:CompanyName"`
-	PhoneNumber  string `xml:"q0:PhoneNumber"`
-	EmailAddress string `xml:"q0:EMailAddress"`
-}
-
-type Dimensions struct {
-	Length int    `xml:"q0:Length"`
-	Width  int    `xml:"q0:Width"`
-	Height int    `xml:"q0:Height"`
-	Units  string `xml:"q0:Units"`
-}
-
-type Payment struct {
-	PaymentType string `xml:"q0:PaymentType"`
-	Payor       Payor  `xml:"q0:Payor"`
-}
-
-type Payor struct {
-	ResponsibleParty ResponsibleParty `xml:"q0:ResponsibleParty"`
-}
-
-type ResponsibleParty struct {
-	AccountNumber string `xml:"q0:AccountNumber"`
-}
-
-type LabelSpecification struct {
-	LabelFormatType string `xml:"q0:LabelFormatType"`
-	ImageType       string `xml:"q0:ImageType"`
-}
-
-type Shipper struct {
-	AccountNumber string  `xml:"q0:AccountNumber"`
-	Contact       Contact `xml:"q0:Contact"`
-	Address       Address `xml:"q0:Address"`
-}
-
-// Reply has common stuff on all responses from FedEx API
-type Reply struct {
-	HighestSeverity string
-	Notifications   []Notification
-	Version         VersionResponse
-	JobID           string `xml:"JobId"`
-}
-
-func (r Reply) Error() error {
-	if r.HighestSeverity == notificationSeveritySuccess {
-		return nil
-	}
-
-	for _, notification := range r.Notifications {
-		if notification.Severity == r.HighestSeverity {
-			return fmt.Errorf("reply got error: %s", notification.Message)
-		}
-	}
-	return fmt.Errorf("reply got status: %s", r.HighestSeverity)
-
-}
-
-// TrackReply : Track reply root (`xml:"Body>TrackReply"`)
-type TrackReply struct {
-	Reply
-	CompletedTrackDetails []CompletedTrackDetail
-}
-
-func (tr *TrackReply) searchDatesOrTimes(dateOrTimeType string) *time.Time {
-	for _, completedTrackDetail := range tr.CompletedTrackDetails {
-		for _, trackDetail := range completedTrackDetail.TrackDetails {
-			for _, dateOrTime := range trackDetail.DatesOrTimes {
-				if dateOrTime.Type == dateOrTimeType {
-					ts := time.Time(dateOrTime.DateOrTimestamp)
-					return &ts
-				}
-			}
-		}
-	}
-
-	return nil
-}
-
-// ActualDelivery returns the first ACTUAL_DELIVERY timestamp
-func (tr *TrackReply) ActualDelivery() *time.Time {
-	return tr.searchDatesOrTimes("ACTUAL_DELIVERY")
-}
-
-// EstimatedDelivery returns the first ESTIMATED_DELIVERY timestamp
-func (tr *TrackReply) EstimatedDelivery() *time.Time {
-	return tr.searchDatesOrTimes("ESTIMATED_DELIVERY")
-}
-
-// Ship returns the first SHIP timestamp
-func (tr *TrackReply) Ship() *time.Time {
-	return tr.searchDatesOrTimes("SHIP")
-}
-
-func (tr *TrackReply) Events() []Event {
-	events := []Event{}
-	for _, completedTrackDetail := range tr.CompletedTrackDetails {
-		for _, trackDetail := range completedTrackDetail.TrackDetails {
-			events = append(events, trackDetail.Events...)
-		}
-	}
-	return events
-}
-
-// ProcessShipReply : Process shipment reply root (`xml:"Body>ProcessShipmentReply"`)
-type ProcessShipmentReply struct {
-	Reply
-	TransactionDetail       TransactionDetail
-	CompletedShipmentDetail CompletedShipmentDetail
-}
-
-// RateReply : Process shipment reply root (`xml:"Body>RateReply"`)
-type RateReply struct {
-	Reply
-	TransactionDetail TransactionDetail
-	RateReplyDetails  []RateReplyDetail
-}
-
-// TotalCost returns the first TotalNetChargeWithDutiesAndTaxes in the reply
-func (rr *RateReply) TotalCost() (Charge, error) {
-	// TotalNetChargeWithDutiesAndTaxes
-	for _, rateReplyDetail := range rr.RateReplyDetails {
-		for _, ratedShipmentDetail := range rateReplyDetail.RatedShipmentDetails {
-			totalNetCharge := ratedShipmentDetail.ShipmentRateDetail.TotalNetChargeWithDutiesAndTaxes
-			if totalNetCharge.Currency != "" && totalNetCharge.Amount != "" {
-				return totalNetCharge, nil
-			}
-		}
-	}
-	return Charge{}, errors.New("no total net charge found on reply")
-}
-
-// CreatePickupReply : CreatePickup reply root (`xml:"Body>CreatePickupReply"`)
-type CreatePickupReply struct {
-	Reply
-	PickupConfirmationNumber string
-	Location                 string
-}
-
-type RateReplyDetail struct {
-	ServiceType                     string
-	ServiceDescription              ServiceDescription
-	PackagingType                   string
-	DestinationAirportID            string `xml:"DestinationAirportId"`
-	IneligibleForMoneyBackGuarantee bool
-	SignatureOption                 string
-	ActualRateType                  string
-	RatedShipmentDetails            []Rating // TODO
-}
-
-type TransactionDetail struct {
-	CustomerTransactionID string `xml:"q0:CustomerTransactionId,omitempty"`
-}
-
-type CompletedShipmentDetail struct {
-	UsDomestic              string
-	CarrierCode             string
-	MasterTrackingId        TrackingID
-	ServiceTypeDescription  string
-	ServiceDescription      ServiceDescription
-	PackagingDescription    string
-	OperationalDetail       OperationalDetail
-	ShipmentRating          Rating
-	CompletedPackageDetails CompletedPackageDetails
-}
-
-type Part struct {
-	DocumentPartSequenceNumber string
-	Image                      []byte
-}
-
-type Label struct {
-	Type                        string
-	ShippingDocumentDisposition string
-	ImageType                   string
-	Resolution                  string
-	CopiesToPrint               string
-	Parts                       []Part
-}
-
-type CompletedPackageDetails struct {
-	SequenceNumber string
-	TrackingIds    []TrackingID
-	Label          Label
-}
-
-type TrackingID struct {
-	TrackingIdType string
-	TrackingNumber string
-}
-
-type Name struct {
-	Type     string
-	Encoding string
-	Value    string
+type Service struct {
+	Type             string
+	Description      string
+	ShortDescription string
 }
 
 type ServiceDescription struct {
@@ -488,88 +394,52 @@ type ServiceDescription struct {
 	AstraDescription string
 }
 
+type ShipmentManifestDetail struct {
+	ManifestReferenceType string `xml:"q0:ManifestReferenceType,omitempty"`
+}
+
+type Shipper struct {
+	AccountNumber string  `xml:"q0:AccountNumber"`
+	Contact       Contact `xml:"q0:Contact"`
+	Address       Address `xml:"q0:Address"`
+}
+
+type SmartPostDetail struct {
+	Indicia              string `xml:"q0:Indicia"`
+	AncillaryEndorsement string `xml:"q0:AncillaryEndorsement"`
+	HubID                string `xml:"q0:HubId"`
+}
+
+type SpecialHandling struct {
+	Type        string
+	Description string
+	PaymentType string
+}
+
+type SpecialServicesRequested struct {
+	SpecialServiceTypes     []string                 `xml:"q0:SpecialServiceTypes,omitempty"`
+	EventNotificationDetail *EventNotificationDetail `xml:"q0:EventNotificationDetail,omitempty"`
+	ReturnShipmentDetail    *ReturnShipmentDetail    `xml:"q0:ReturnShipmentDetail,omitempty"`
+}
+
+type StatusDetail struct {
+	CreationTime     Timestamp
+	Code             string
+	Description      string
+	Location         Address
+	AncillaryDetails []AncillaryDetail
+}
+
+type StringBarcode struct {
+	Type  string
+	Value string
+}
+
 type Surcharge struct {
 	SurchargeType string
 	Level         string
 	Description   string
 	Amount        Charge
-}
-
-type OperationalDetail struct {
-	OriginLocationNumber            string
-	DestinationLocationNumber       string
-	TransitTime                     string
-	IneligibleForMoneyBackGuarantee string
-	DeliveryEligibilities           string
-	ServiceCode                     string
-	PackagingCode                   string
-}
-
-type RatedShipmentDetail struct {
-	EffectiveNetDiscount Charge
-	ShipmentRateDetail   RateDetail
-	RatedPackages        []RatedPackage
-}
-
-type Rating struct {
-	ActualRateType       string
-	GroupNumber          string
-	EffectiveNetDiscount Charge
-
-	// For the shipping service, the rate details is an array, but for the rate service, it is not
-	ShipmentRateDetails []RateDetail
-	ShipmentRateDetail  RateDetail
-
-	RatedPackages []RatedPackage
-}
-
-type Charge struct {
-	Currency string
-	Amount   string
-}
-
-type RatedPackage struct {
-	GroupNumber          string
-	EffectiveNetDiscount Charge
-	PackageRateDetail    RateDetail
-}
-
-type RateDetail struct {
-	RateType                         string
-	RateZone                         string
-	RatedWeightMethod                string
-	DimDivisor                       string
-	FuelSurchargePercent             string
-	TotalBillingWeight               Weight
-	TotalBaseCharge                  Charge
-	TotalFreightDiscounts            Charge
-	TotalNetFreight                  Charge
-	TotalSurcharges                  Charge
-	TotalNetFedExCharge              Charge
-	TotalTaxes                       Charge
-	TotalNetCharge                   Charge
-	NetCharge                        Charge
-	TotalRebates                     Charge
-	TotalDutiesAndTaxes              Charge
-	TotalAncillaryFeesAndTaxes       Charge
-	TotalDutiesTaxesAndFees          Charge
-	TotalNetChargeWithDutiesAndTaxes Charge
-	Surcharges                       []Surcharge
-}
-
-type VersionResponse struct {
-	ServiceID    string `xml:"ServiceId"`
-	Major        int
-	Intermediate int
-	Minor        int
-}
-
-type CompletedTrackDetail struct {
-	HighestSeverity  string
-	Notifications    []Notification
-	DuplicateWaybill bool
-	MoreData         bool
-	TrackDetails     []TrackDetail
 }
 
 type TrackDetail struct {
@@ -631,104 +501,16 @@ type TrackDetail struct {
 	Events                                 []Event
 }
 
-type DateOrTimestamp struct {
-	Type            string
-	DateOrTimestamp Timestamp
+type TrackingID struct {
+	TrackingIdType string
+	TrackingNumber string
 }
 
-type AdvanceNotificationDetail struct {
-	EstimatedTimeOfArrival Timestamp
-	Reason                 string
-	Status                 string
-	StatusDescription      string
-	StatusTime             Timestamp
+type TransactionDetail struct {
+	CustomerTransactionID string `xml:"q0:CustomerTransactionId,omitempty"`
 }
 
-type ContentRecord struct {
-	PartNumber       string
-	ItemNumber       string
-	ReceivedQuantity int
-	Description      string
-}
-
-type ContactAndAddress struct {
-	Contact Contact `xml:"q0:Contact"`
-	Address Address `xml:"q0:Address"`
-}
-
-type Reconciliation struct {
-	Status      string
-	Description string
-}
-
-type InformationNoteDetail struct {
-	Code        string
-	Description string
-}
-
-type StringBarcode struct {
-	Type  string
-	Value string
-}
-
-type Notification struct {
-	Severity         string
-	Source           string
-	Code             string
-	Message          string
-	LocalizedMessage string
-}
-
-type StatusDetail struct {
-	CreationTime     Timestamp
-	Code             string
-	Description      string
-	Location         Address
-	AncillaryDetails []AncillaryDetail
-}
-
-type Address struct {
-	StreetLines         []string `xml:"q0:StreetLines"`
-	City                string   `xml:"q0:City"`
-	StateOrProvinceCode string   `xml:"q0:StateOrProvinceCode"`
-	PostalCode          string   `xml:"q0:PostalCode"`
-	CountryCode         string   `xml:"q0:CountryCode"`
-	// CountryName         string   `xml:"q0:CountryName"`
-	Residential Bool `xml:"q0:Residential"`
-}
-
-type AncillaryDetail struct {
-	Reason            string
-	ReasonDescription string
-}
-
-type OtherIdentifier struct {
-	PackageIdentifier Identifier
-}
-
-type Service struct {
-	Type             string
-	Description      string
-	ShortDescription string
-}
-
-type Identifier struct {
-	Type  string
-	Value string
-}
-
-type SpecialHandling struct {
-	Type        string
-	Description string
-	PaymentType string
-}
-
-type Event struct {
-	Timestamp                  time.Time
-	EventType                  string
-	EventDescription           string
-	StatusExceptionCode        string
-	StatusExceptionDescription string
-	Address                    Address
-	ArrivalLocation            string
+type Weight struct {
+	Units string  `xml:"q0:Units"`
+	Value float64 `xml:"q0:Value"`
 }
